Modules::String Move implementaion from string class to separate files
